refactor: Update payment arguments

+++ api/calc/internal/domain/payment/entity.go
@@ -9,7 +9,7 @@ type Payment struct {
 	ID        string    `firestore:"id"`         // UUID
 	Name      string    `firestore:"name"`       // 支払い名
 	Currency  string    `firestore:"currency"`   // 通貨
-	Total     int       `firestore:"total"`      // 合計金額
+	Total     float64   `firestore:"total"`      // 合計金額
 	Payers    []*Payer  `firestore:"payers"`     // 支払い者情報一覧
 	Tags      []string  `firestore:"tags"`       // タグ
 	Comment   string    `firestore:"comment"`    // コメント
@@ -21,6 +21,6 @@ type Payment struct {
 
 // Payer - 支払い者 エンティティ
 type Payer struct {
-	ID     string `firestore:"id"`     // ユーザーID
-	Amount int    `firestore:"amount"` // 支払い金額
+	ID     string  `firestore:"id"`     // ユーザーID
+	Amount float64 `firestore:"amount"` // 支払い金額
 }
